fix : 대각선 이동 시 상대 말을 먹지 못하는 버그 수정

+++ src/main/java/chess/domain/piece/Direction.java
@@ -81,7 +81,7 @@ public enum Direction {
         return yDegree;
     }
 
-    public int getUnit() {
-        return (int) Math.sqrt(xDegree * xDegree + yDegree * yDegree);
+    public double getUnit() {
+        return Math.sqrt(xDegree * xDegree + yDegree * yDegree);
     }
 }
